+++ packages/bundler-plugin-core/src/index.ts
@@ -14,6 +14,7 @@ import {
 import { red } from "./utils/logging.ts";
 import { normalizePath } from "./utils/normalizePath.ts";
 import { bundleAnalysisPluginFactory } from "./bundle-analysis/bundleAnalysisPluginFactory.ts";
+import { normalizeOptions } from "./utils/normalizeOptions.ts";
 
 const NODE_VERSION_RANGE = ">=18.18.0";
 
@@ -27,6 +28,15 @@ function codecovUnpluginFactory({
   return createUnplugin<Options, true>((userOptions, unpluginMetaContext) => {
     const plugins: UnpluginOptions[] = [];
 
+    const normalizedOptions = normalizeOptions(userOptions);
+
+    if (!normalizedOptions.success) {
+      for (const error of normalizedOptions.errors) {
+        red(error);
+      }
+      return [];
+    }
+
     if (!satisfies(process.version, NODE_VERSION_RANGE)) {
       red(
         `Codecov ${unpluginMetaContext.framework} bundler plugin requires Node.js ${NODE_VERSION_RANGE}. You are using Node.js ${process.version}. Please upgrade your Node.js version.`,
@@ -35,10 +45,11 @@ function codecovUnpluginFactory({
       return plugins;
     }
 
-    if (userOptions?.enableBundleAnalysis) {
+    const options = normalizedOptions.options;
+    if (options?.enableBundleAnalysis) {
       plugins.push(
         bundleAnalysisPluginFactory({
-          userOptions,
+          options,
           bundleAnalysisUploadPlugin,
         }),
       );