+++ packages/bundler-plugin-core/src/utils/providers/Vercel.ts
@@ -1,3 +1,6 @@
+// VERCEL ENV DOCS
+// https://vercel.com/docs/projects/environment-variables/system-environment-variables
+
 import {
   type ProviderEnvs,
   type ProviderServiceParams,
@@ -18,9 +21,9 @@ function _getBuildURL(): string {
 }
 
 function _getBranch(inputs: ProviderUtilInputs): string {
-  const { args } = inputs;
+  const { args, envs } = inputs;
 
-  return args?.branch ?? "";
+  return args?.branch ?? envs?.VERCEL_GIT_COMMIT_REF ?? "";
 }
 
 function _getJob(): string {
@@ -48,7 +51,15 @@ function _getSHA(inputs: ProviderUtilInputs): string {
 function _getSlug(inputs: ProviderUtilInputs): string {
   const { args, envs } = inputs;
   if (args?.slug && args?.slug !== "") return args?.slug;
-  return envs?.VERCEL_GIT_REPO_SLUG ?? "";
+  const owner = envs?.VERCEL_GIT_REPO_OWNER ?? "";
+  const repo = envs?.VERCEL_GIT_REPO_SLUG ?? "";
+
+  let slug = "";
+  if (owner && repo) {
+    slug = `${owner}/${repo}`;
+  }
+
+  return slug;
 }
 
 // eslint-disable-next-line @typescript-eslint/require-await