+++ packages/vite-plugin/src/vite-bundle-analysis/viteBundleAnalysisPlugin.ts
@@ -12,7 +12,7 @@ const PLUGIN_NAME = "codecov-vite-bundle-analysis-plugin";
 
 export const viteBundleAnalysisPlugin: BundleAnalysisUploadPlugin = ({
   output,
-  userOptions,
+  options: userOptions,
 }) => ({
   version: "1",
   name: PLUGIN_NAME,
@@ -149,7 +149,7 @@ export const viteBundleAnalysisPlugin: BundleAnalysisUploadPlugin = ({
       output.outputPath = options.dir ?? "";
 
       // only output file if running dry run
-      if (userOptions?.dryRun) {
+      if (userOptions.dryRun) {
         this.emitFile({
           type: "asset",
           fileName: "codecov-bundle-stats.json",