+++ packages/webpack-plugin/src/webpack-bundle-analysis/__tests__/webpackBundleAnalysisPlugin.test.ts
@@ -7,8 +7,12 @@ describe("webpackBundleAnalysisPlugin", () => {
         output: {
           bundleName: "test",
         },
-        userOptions: {
+        options: {
           bundleName: "test",
+          apiUrl: "http://localhost",
+          dryRun: true,
+          enableBundleAnalysis: true,
+          retryCount: 1,
         },
       });